+++ device-bootloaders/src/ServiceProcessorBL2/driver/pmic_controller.c
@@ -182,7 +182,7 @@ int wait_pmic_ready(void)
 {
 #if !FAST_BOOT
     uint64_t elapsed_ms;
-    const uint64_t timeout_ms = 3;
+    const uint64_t timeout_ms = 5;
     const uint64_t start_ticks = timer_get_ticks_count();
     do
     {
@@ -195,7 +195,7 @@ int wait_pmic_ready(void)
         elapsed_ms = timer_convert_ticks_to_ms(timer_get_ticks_count() - start_ticks);
     } while (elapsed_ms < timeout_ms);
 
-    Log_Write(LOG_LEVEL_ERROR, "wait_pmic_ready timed out after %lu ms\n", timeout_ms);
+    Log_Write(LOG_LEVEL_WARNING, "wait_pmic_ready timed out after %lu ms\n", timeout_ms);
     return -1;
 #else
     return 0;