+++ device-bootloaders/src/ServiceProcessorBL2/services/thermal_power_monitor.c
@@ -303,18 +303,11 @@ static void pwr_svc_set_module_voltage(uint16_t tag, uint64_t req_start_time, vo
     struct device_mgmt_set_module_voltage_rsp_t dm_rsp;
     int32_t status;
 
-    status = pmic_set_voltage(set_voltage_cmd->type, set_voltage_cmd->value);
-    if (STATUS_SUCCESS == status)
+    status = Thermal_Pwr_Mgmt_Set_Validate_Voltage(set_voltage_cmd->type, set_voltage_cmd->value);
+    if (STATUS_SUCCESS != status)
     {
-#if !(FAST_BOOT || TEST_FRAMEWORK)
-        status = check_voltage_stability(set_voltage_cmd->type, set_voltage_cmd->value);
-#else
-        status = STATUS_SUCCESS;
-#endif
-    }
-    else
-    {
-        Log_Write(LOG_LEVEL_ERROR, " thermal pwr mgmt error: set_module_voltage()\r\n");
+        Log_Write(LOG_LEVEL_ERROR,
+                  " thermal pwr mgmt error %d: Thermal_Pwr_Mgmt_Set_Validate_Voltage \r\n", status);
     }
 
     FILL_RSP_HEADER(dm_rsp, tag, DM_CMD_SET_MODULE_VOLTAGE,